Adding an ECDSA file with vrs encoding.
The file
ecdsa_normalized_secp256k1_keccak256_ethereum_main.json
contains ECDSA test vectors, where the signatures are 65 bytes long. The first byte of the signature is a value v, which is equal to 37 + recovery_id, where the recovery_id is a value that allows to reconstruct the point R from the signature value r. Signatures are normalized, which means that a signature is only considered valid if 0 < s < n/2. Public keys are not compressed.
The test vectors are meant to verify ECDSA verification through public key recovery.  Many of the test vectors have been constructed such that internal values computed during the public key reconstruction are edge cases and hence may uncover arithmetic overflows and underflows during the computation.

The file is not tested against any third party implementation. Additionally, I don't have a comprehensive standard that properly defines which signatures are valid and which signatures should be rejected. For example there does not seem to be consensus, whether signatures with recovery_id 3,4 are valid or should be rejected