Adding different ECDSA versions
Adding different versions for ECDSA depending on whether signatures need to be normalized or not.

The main goal here is to determine which version are used by common protocols.

There is a small format change. The test vectors now contain a description. This description contains a short definition of the formats used for the test vectors, with the goal to avoid confusion caused by wrong expectations.