Adding test status
This change add the status of the test vectors.

All tables are available in both JSON and YAML format.

all_sets.json  contains a description of all the test vector files. The description contains the current test status against some cryptographic libraries. One of the goals of this file is to add some clarity what algorithms are implemented by different libraries, since libraries sometimes do not specify all algorithm parameters.
reference.json contains the bibliography in JSON format.
index.json is an index that lists the references together with the files based on this reference.