Removing test vectors for ECDH in raw format.
This format is used BigIntegers represented as integers. Many JSON parsers have problems with BigIntegers. The test vectors itself are not very useful, since crypto libraries generally have some support for at least one of the other key formats (PEM, PKCS8, etc)

For testing ECDH implementations it is quite important that the test vectors support the key format of the library to tests, so that the libraries can be checked for things like invalid curve attacks.