New test vectors for BIP-340.
Old versions of BIP-340 specified that the signed data is 32-bytes long.
While this restriction has been removed, there still exist libraries that
restrict the size of the input to 32-bytes.
Most of the test vectors have been changed to have a 32-byte message.
This allows to test libraries with restrictions on the message length.